requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.1.0
PyYAML==6.0.1
django-extensions==3.2.3
coreapi==2.3.3
coreschema==0.0.4
itypes==1.2.0
uritemplate==4.1.1
drf-yasg==1.21.7
setuptools==69.0.2